adding activation & output functions.
I really don't know when I should be using a reference or when I should
be using a pointer or when I should be using the actual symbol.

is c++ pass by reference or pass by value? will it copy the whole thing
if I don't do a reference?

I wonder if I could get gpt4 to read my commits and my commit messages
and give me an answer...