performing classification on the MNIST test data.
this is just to see if things are working, and see what the random
weights think about the MNIST data.

there's about 150 lines of code now in a language I don't know (I guess
I'm writing javascript in c++ at this point) so there is likely a bug in
there somewhere.  or just something silly.

the grown up SWE in me says I should write tests next.
but nah...