using pointers to train the model.
testing by comparing a sample of the weights after one round of training
on a 100-long subset, still using the static unique weights from before.
the output code from last commit is still there.
running with just the old code has it around 26s for an epoch on 100
samples.
running with just the pointers code is around 0.1s.  so two orders of
magnitude speedup.

a lot of that speed up is probably just reduced computations rather than
memory issues, though, the new training function computes and saves all
the outputs, the dot products, and the deriv of the activation functions
evaluated on the outputs (the primed outputs, I guess?) before it goes
into any for loop and uses those results.  so the old code repeated the
same computations a few hundred times, nevermind any possible copying
across in memory.