some tests on the autodiff, plus comparing xor
a multilayer "percepton" was implemented in both models, tried to learn
XOR using it and compared the gradients.

the tests could be better but they found a bug that was keeping the
autodiff from working.

the gradients are different on the inner weights but the model still
seems to learn.