New autodiff using double* and matrices.
The resulting autodiff is much faster when its being used to implement a
forward feed neural network (ffnn).  Also much nicer to program with.

There are still many slow spots (like the copying from valarray<double>
to double* in the Col).  The calc for the gradient of a matrix doesn't
really make sense, either (it is just the calc I need for the ffnn).

Primary reason for this is to make it easier to use cuda for the matrix
product.  Also to make it easier to write a CNN.

Scikit-learn's implementation is still 100x faster.