convolution via matrix-vector product (wip).
the convolution operation works but much to clean up.
unsure if the unpacking of the kernel into the bigger matrix is
efficient - not sure if cuda is the best place for all that integer
arithmetic and bool checking.  also there is shared memory with the
kernel there that I could be using.  similar concerns with the padding.
the gradient is still not implemented.