autodiff for conv operation + tests.
removing some memory leaks in the test code as well.
the key code for the convolution autodiff is in the doKernelRoll
cuda kernel code.  The idea here is if we "unroll" the kernel into
a matrix to do convolution as a matrix-vector product, this rolls
back the gradient back up into the smaller kernel.

The algorithm is sorta funny.  it runs on every column in the gradient
matrix, and then when it finds a column, each kernel will sum up the
bits in each row.

this means that most cuda cores are going to be idle, because they won't
have something to do, so we are blocking a lot of cuda resources for
this.

the test suite could really use some structure.  might need to look into
moving to cmake and using a unit test library.