Adding a ConcatCol layer.
Turns several column vectors into on giant column vector.
As explained in the comments this is a bit stupid as its just moving
cuda device memory blocks around.

A slightly less daft idea would be to have ConcatCol re-allocate the
memory if each of its members so that there is one big block, and then
have its d_value and d_grad point to those blocks.  But your author
doesn't know how to do that [yet] and this was the first idea that came
to my mind.