ConcatCol reuses its d_value for its members.
The note we wrote about allocated a big chunk of memory for the d_value
and having each component col just use a block in that makes sense and
it isn't hard to implement.  So we did.

However, for pushGrad things are a bit more difficult, since each
member's pushGrad wants to cudaFree its d_seed.  So if we pass in
something like d_seed + memIndex, someone else will try to cudaFree it,
causing some problems.  So that's still doing the stupid copying.

This also shows a problem with my design of building up a DAG for the
computation that has different nodes responsible for cleaning up another
nodes memory.  May need to re-think the whole architecture here.