starting on autograd.
learned about cublasSetPointerMode - apparently it's needed to use alpha
on the device.  I'm only doing this when computing the deriv of an inner
product, so I switch it back.  unsure if this is good practice or not.

TODO I will need another innerproduct test, taking \del/\del(x) (x*y).