starting Matrix-vector mult, some prelim refactors
using a switch instead of ifs.
is there exhaustive pattern matching in switches in c++?

One issue here is where should the values of the matrix live?  its fine
for them to live in _working, but if the values of the matrix are
themselves the result of another operation then it should be in _result.
this gets a bit tricky.

it's probably better to have an InputMatrix and a MatrixMult function
that is a BinaryOperation.  I might be able to make it generic so its
matrix-matrix not just matrix-column.  but I don't need that generality
right now so I won't go that far, but will proceed with this refactor
before continuing.