FFNN implementation in daft + error handling
1. cublas error handling via a macro
2. cudas error handling via a macro
3. fixed a bug in the matrix-matrix product multiplication that was hard
   to catch and causing daft ffnn's gradients to blow up and diverge

1 and 2 are stock code you can get from stackoverflow, google, llms.
3 was hard to find, and I put in the error codes (And a bunch of device
syncs) to try to find the problem.

I really didn't get anywhere until I outputed the weights and the
activations for layer 1 at each step.  at that point it was clear the
activations were different.  I used claude at this point (llm) to get
some help and it reminded me / pointed out that I'm using gemm for the
matrix mult in daft but gemv in silly.  The real kicker is that I didn't
do the transpose right.  since I do things in column-order and cublas
does it in row-order I have to transpose, which I didn't do.
transposing fixed the problem.

this is actually slightly slower than silly, but its probably due to the
extxra device syncs everywhere...