implementing concat.
concat assumes that all its targets are a continuous block in memory
and then for the gradient it slices up the seed and pushes it down.

it could copy them all into a new block but that would be really slow.
a smarter compile() function might do that automatically.
this compile() just makes sure that the _result of the concat is the
_result of its first element and hopes you got size right.