batch processing for matrix multiplication.
tests for single op pass with this new code, but I disabled the tests
for the grad.  TODO write tests that actually test batch processing!

NOTE: InputMatrix is now special - it doesn't expect to be batched.
think about weights.  this may or may not be the right way to handle
this..