batchProcessing for leakyReLU
includes tests and tests for the grad.
this means updating the grad for InputColumn.
(note that for the grad for InputMatrix we need to be adding all results
 into the same matrix, rather than batches, because we don't batch input
 matrices as those should be weights.)

there is no batchedSaxpy for cublas, so we wrote a kernel to batch add.
now, for this case the memory is all layed out in order.  so we could
just do one big giant vector add of the seed and the grad.
we could still have used Saxpy even!

but I split it up for clarity's sake.  I might go back and do a
cublasSaxpy later with n = gradSize * batchSize. that should be a wee
bit faster as it would have fewer checks and more kernels in the same
branch path.

but for now I'm going for clarity of code and what I'm doing.